src/app/shared/components/focus-timer/focus-timer.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { FocusTimerComponent } from './focus-timer.component';

describe('Component: FocusTimer', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [FocusTimerComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([FocusTimerComponent],
      (component: FocusTimerComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(FocusTimerComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(FocusTimerComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-focus-timer></app-focus-timer>
  `,
  directives: [FocusTimerComponent]
})
class FocusTimerComponentTestController {
}

